+++ R/ddfilter_speed.R
@@ -14,15 +14,14 @@
 #' @param vmax A numeric value specifying a threshold of speed from a previous and/or to a subsequent fix. 
 #' Default is 8.9km/h. 
 #' If this value is unknown, it can be estimated from \emph{sdata} using the function \code{\link{vmax}}.
-#' @param method An integer (1 or 2) specifying how locations should be filtered with \emph{vmax}. 
-#' Default is 1 (both way) and removes a location if the speed from a previous AND to a subsequent location exceeds \emph{vmax}. 
-#' Select 2 (one way) to remove a location if the speed from a previous OR to a subsequent location exceeds \emph{vmax}. 
+#' @param method An integer (1 or 2) specifying how locations should be filtered. 
+#' Default is 1 and removes a location if the speed from a previous AND to a subsequent location exceeds \emph{vmax}. 
+#' Select 2 to remove a location if the speed from a previous OR to a subsequent location exceeds \emph{vmax}. 
 #' For the latter, the filter examines successive suspect locations (i.e. the speed from a previous and/or to a subsequent location exceeds \emph{vmax}) 
-#' and retain one location that is associated with the minimum speed from a previous and/or to a subsequent location.
+#' and retain one location that is associated with the minimum speed from a previous and to a subsequent location.
 #' @importFrom dplyr bind_rows
 #' @export
 #' @details This function removes locations if the speed from a previous and/or to a subsequent location exceeds a given threshold speed. 
-#' When method = 2 is selected (one way), 
 #' If \emph{vmax} is unknown, it can be estimated using the function \code{\link{vmax}}.
 #' @return The input data is returned without locations identified by this filter. 
 #' The following columns are added: "pTime", "sTime", "pDist", "sDist", "pSpeed", "sSpeed". 
@@ -82,32 +81,33 @@ ddfilter_speed <- function (sdata, vmax = 8.9, method = 1){
     } else if (method == 2){
       
       # Remove locations at which the speed from a previous OR to a subsequent location exceeds maximum linear traveling speed (Vmax)
+      sdata$row <- seq(1, nrow(sdata))
       sdata1 <- with(sdata, sdata[(!is.na(pSpeed) & pSpeed > vmax) | (!is.na(sSpeed) & sSpeed > vmax),])
- 
+
       while(nrow(sdata1) > 0){
         # group successive locations to assess
-        index <- 0; g <- rep(0, nrow(sdata1))
+        index <- 1; g <- rep(1, nrow(sdata1))
         for(i in 2:nrow(sdata1)){
-          if(sdata1[i, 'pSpeed'] != sdata1[i-1, 'sSpeed']){
+          if(is.na(sdata1[i, 'pSpeed']) | sdata1[i, 'pSpeed'] != sdata1[i-1, 'sSpeed'] | is.na(sdata1[i-1, 'sSpeed'])){
             index <- index + 1
-          } 
+          }
           g[i] <- index
         }
         sdata1$group <- g
-        
+
         # Remove/retain locations based on the mean speed from a previous and to a subsequent location
-        sdata1$meanSpeed <- rowMeans(sdata1[, c('pSpeed', 'sSpeed')], na.rm = TRUE)
+        rm_row <- rep(0, length(unique(sdata1$group)))
         for(i in unique(sdata1$group)){
-          if(length(sdata1[sdata1$group %in% i, 'group']) > 1){
-            max_meanSpd <- max(with(sdata1, sdata1[group %in% i, 'meanSpeed']))
-            rm_row <- with(sdata1, sdata1[group %in% i & meanSpeed %in% max_meanSpd,])
-            sdata <- with(sdata, sdata[!(id %in% unique(rm_row$id) & DateTime %in% rm_row$DateTime),])
-          }
-        }
-        
+          temp_df <- sdata1[sdata1$group %in% i, ]
+          mean_spd <- with(temp_df, rowMeans(cbind(pSpeed, sSpeed), na.rm = TRUE))
+          rm_row[i] <- with(temp_df, temp_df[which.max(mean_spd), 'row'])
+         }
+
+        sdata <- sdata[!(sdata$row %in% rm_row),]
         sdata <- track_param(sdata, param = 'speed')
         sdata1 <- with(sdata, sdata[(!is.na(pSpeed) & pSpeed > vmax) | (!is.na(sSpeed) & sSpeed > vmax),])
       }
+      sdata$row <- NULL
     }
   }